requirements.txt
pathway>=0.14.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.5.0
python-multipart>=0.0.9
aiofiles>=23.2.1
